fix: 수정사항반영

See merge request pjt/bsp/sms/sms-wells-back-app!6394

+++ src/main/java/com/kyowon/sms/wells/web/competence/evaluate/dvo/WpsdElvDetailDvo.java
@@ -13,8 +13,8 @@ public class WpsdElvDetailDvo {
     private String evlAtcDvCd;        /* 평가항목구분코드 */
     private String evlPdDvCd;         /* 평가상품구분코드 */
     private double evlBaseUnitN;     /* 평가기준단위수 */
-    private Long cvtBasePc;           /* 환산기준점수 */
-    private Long trgBasePc;           /* 목표기준점수 */
+    private double cvtBasePc;           /* 환산기준점수 */
+    private double trgBasePc;           /* 목표기준점수 */
     private String ctstGrpCd;         /* 경진그룹코드 */
     private String dtaDlYn;           /* 데이터삭제여부 */
     private String rowState;
